added dark mode to own file added text to togglealgorithim in successfully changing text when changing mode
